Added updating information about agents and tracking of agent availability (e.g., on duty, off duty, on leave).

+++ src/main/java/com/example/p8t2adrianaflorinroxana/repository/AgentRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface AgentRepository extends JpaRepository<Agents, Integer> {
+public interface AgentRepository extends JpaRepository<Agents, Long> {
     List<Agents> findByStation_Id(long stationId);
 
     void deleteById(long id);
